R 10 MainWindowInicialisationFlag refactor
