+++ src/Python/FrontEnd/MainWindow/Abstract/MainWindowAbstract.py
@@ -6,6 +6,10 @@ from Python.BaseClass.Logger.Logger import Loger
 class MainWindowAbstract(QMainWindow, Loger):
     zoomInterface = None
 
+    menu = None
+
+    cameraView = None
+
     def qActionCreate(self, name: str, triggerFun, checkable=False) -> QAction:
         qAction = QAction(name, self)
         qAction.setCheckable(checkable)
