+++ src/Python/StartUP/12-2024-12.log
@@ -1,0 +1,31 @@
+[2024-12-12 13:22:12.412165] - [m] - [mainNoTryCahch] [log] [["res PyQt5.QtCore.QSize(1920, 1080) <class 'PyQt5.QtCore.QSize'>"]]
+[2024-12-12 13:22:12.618315] - [MainWindow] - [_logImportant] [log] [['####################################################################################################']]
+[2024-12-12 13:22:12.621210] - [MainWindow] - [_logImportant] [log] [['######################################### Program Starting ##########################################']]
+[2024-12-12 13:22:12.623747] - [MainWindow] - [_logImportant] [log] [['####################################################################################################']]
+[2024-12-12 13:22:12.760387] - [Camera] - [__init__] [log] [['Connection to Camera from Produce 0:']]
+[2024-12-12 13:22:12.766683] - [Camera] - [setVideoFormat] [log] [['ustawienie formatu Wideo res: None']]
+[2024-12-12 13:22:12.770797] - [Camera] - [StartLive] [ERROR] [[IndexError('list index out of range')]]
+[2024-12-12 13:22:12.778303] - [Camera] - [StartLive] [ERROR] [[['logError', 'StartLive', '__init__', '__init__', '__init__', '__init__', '__init__', '__init__', '__init__', '__init__', '__init__', '__init__', 'mainNoTryCahch', 'main', '<module>', 'execfile', '_exec', 'run', 'main', '<module>']]]
+[2024-12-12 13:22:12.782304] - [Camera] - [StartLive] [log] [['no window to close?']]
+[2024-12-12 13:22:12.786305] - [Camera] - [__init__] [log] [[(0, 0, 0)]]
+[2024-12-12 13:22:13.727318] - [Camera] - [__init__] [log] [[ValueError('NULL pointer access')]]
+[2024-12-12 13:22:14.518877] - [ManipulatorInterfere] - [resolveManipulator] [log] [[FileNotFoundError("Could not find module 'C:\\Program Files\\PollsampleX\\Config\\Dlls\\ACSCL_x64.dll' (or one of its dependencies). Try using the full path with constructor syntax.")]]
+[2024-12-12 13:22:14.523168] - [AbstractManipulator] - [getCurrentPosition] [Warning] [['Abstract Methode']]
+[2024-12-12 13:22:14.602773] - [StandaManipulator] - [enter] [log] [['Device ID 1']]
+[2024-12-12 13:22:16.708063] - [StandaManipulator] - [__testInfo] [log] [['Result: -4']]
+[2024-12-12 13:22:18.818747] - [StandaManipulator] - [__init__] [ERROR] [['Error reading current position']]
+[2024-12-12 13:22:18.826383] - [StandaManipulator] - [__init__] [ERROR] [[['logError', '__init__', 'resolveManipulator', '__init__', '__init__', '__init__', '__init__', '__init__', '__init__', 'mainNoTryCahch', 'main', '<module>', 'execfile', '_exec', 'run', 'main', '<module>']]]
+[2024-12-12 13:22:20.930908] - [StandaManipulator] - [__init__] [log] [['Standa Resived Position: -4']]
+[2024-12-12 13:22:20.934923] - [StandaManipulator] - [__init__] [ERROR] [[FileNotFoundError(2, 'No such file or directory')]]
+[2024-12-12 13:22:20.942435] - [StandaManipulator] - [__init__] [ERROR] [[['logError', '__init__', 'resolveManipulator', '__init__', '__init__', '__init__', '__init__', '__init__', '__init__', 'mainNoTryCahch', 'main', '<module>', 'execfile', '_exec', 'run', 'main', '<module>']]]
+[2024-12-12 13:22:26.893442] - [StandaManipulator] - [enter] [log] [['Device ID 2']]
+[2024-12-12 13:22:34.059586] - [StandaManipulator] - [__testInfo] [log] [['Result: -4']]
+[2024-12-12 13:22:41.155415] - [StandaManipulator] - [__init__] [ERROR] [['Error reading current position']]
+[2024-12-12 13:22:41.162395] - [StandaManipulator] - [__init__] [ERROR] [[['logError', '__init__', 'resolveManipulator', '__init__', '__init__', '__init__', '__init__', '__init__', '__init__', 'mainNoTryCahch', 'main', '<module>', 'execfile', '_exec', 'run', 'main', '<module>']]]
+[2024-12-12 13:22:48.367447] - [StandaManipulator] - [__init__] [log] [['Standa Resived Position: -4']]
+[2024-12-12 13:22:48.371590] - [StandaManipulator] - [__init__] [ERROR] [[FileNotFoundError(2, 'No such file or directory')]]
+[2024-12-12 13:22:48.378592] - [StandaManipulator] - [__init__] [ERROR] [[['logError', '__init__', 'resolveManipulator', '__init__', '__init__', '__init__', '__init__', '__init__', '__init__', 'mainNoTryCahch', 'main', '<module>', 'execfile', '_exec', 'run', 'main', '<module>']]]
+[2024-12-12 13:22:48.394728] - [StandaManipulator] - [__goto] [log] [[0.0, 0, 0]]
+[2024-12-12 13:22:57.918983] - [MainWindow] - [createZoom] [log] [['readed zoom 3']]
+[2024-12-12 13:22:57.932398] - [MainWindow] - [__toggle] [log] [[[60.0, 90.0, 60.0, 90.0, 'pole 0']]]
+[2024-12-12 13:23:20.333358] - [AbstractManipulator] - [close] [Warning] [['Abstract Methode']]
